+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiActions.kt
@@ -1,8 +1,11 @@
 package org.simple.clinic.home.overdue.search
 
+import androidx.paging.PagingData
+import org.simple.clinic.home.overdue.OverdueAppointment
 import java.util.UUID
 
 interface OverdueSearchUiActions {
   fun openPatientSummaryScreen(patientUuid: UUID)
   fun openContactPatientSheet(patientUuid: UUID)
+  fun showOverdueSearchResults(searchResults: PagingData<OverdueAppointment>)
 }