+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Search overdue patients, when the search history item is clicked
 - Prefill search query when the search history item is clicked
 - Bump coroutines to v1.6.3
+- Implement view effects to show overdue search results
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300